sibal

+++ workspace_java/JavaBasic/src/사/메서드/Test_2_8.java
@@ -4,7 +4,7 @@ public class Test_2_8 {
     public static void main(String[] args) {
         System.out.println(thr(10,20,30));
     }
-    public static int thr(int a, int b, int c){
-        return (a + b + c) / 3 ;
+    public static double thr(int a, int b, int c){
+        return (a + b + c) / 3.0 ;
     }
 }
